VideoAnalytics: Memory bloat observed during Six inference Demo.
Ingestion rate is high in comparsion to analytics Algo processing speed.
Hence with a huge watermark value for Analytics container, frames seem to be
accumulating in RAM and bloating the memory

Reducing that high water mark(HWM) value alleviates the RAM usage.
The README is altered to reflect the same.

Signed-off-by: Lalatendu das <lalatendu.das@intel.com>